pyproject.toml
[project]
name = "browser-agent"
version = "0.1.0"
description = "Generic browser automation agent with self-correction (Playwright + Claude)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "Kevin Wei", email = "weisl@nlg.csie.ntu.edu.tw"}]

dependencies = [
  "playwright>=1.49",
  "anthropic>=0.40",
  "openai>=1.50",
  "httpx>=0.27",
  "tenacity>=8.5",
  "asyncpg>=0.30",  # selector cache (Postgres-backed)
  "python-dotenv>=1.0",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/workers/browser", "src/shared", "src/server"]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]
